Fixed various command queue locking issues.
Fixed YouTube player widget event handling such that only responded events cause dequeuing.  Fixed an issue with invalid commands causing lockup due to not being responded to.  Fixed an issue with the SoundCloud widget not responding to consecutive pause commands due to state change not triggering.  Fixed an issue with the YouTube widget not responding to consecutive play/pause commands due to state change not triggering.  Fixed an issue wherein if no player has been selected and a command is issued, the command queue would lock.